COG-30: BaseTokenStoreLiquibaseTest is extracted
